apps/api/pyproject.toml
[project]
name = "money-api"
version = "0.1.0"
description = "Money Tracking - backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.3",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "cryptography>=44.0.0",
    "argon2-cffi>=23.1.0",
    "python-multipart>=0.0.20",
    "apscheduler>=3.10.4",
    "email-validator>=2.2.0",
    # LLM / Agent
    "langchain-core>=0.3.30",
    "langchain-ollama>=0.2.2",
    "langgraph>=0.2.60",
    "langfuse>=2.60.0",
    # Gmail
    "google-api-python-client>=2.155.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = ["E501"]
